Improved support for Arduino UNO & Example.
Changes headers from <WProgram.h> to <Arduino.h> for compatibility with
the Arduino IDE, does not change anything for the Teensy.

Adds an example which succesfully received packets when used with an
Arduino UNO. Main problem was that the pin number is not identical
to the interrupt number on this platform.