setLNA() had incorrect type on LnaZin
LnaZin was specified as bool. It needs to be 8 bit in order to have the correct bit field values for setting 200 ohm, which is recommended value in the datasheet and what is used in the example sketch. Because the type was wrong it would set the value to 50 ohm even when specifying 200.