Changed second call to pinMode() to digitalWrite()
Minor correction. The bareRFM69() constructor sets the cs_pin to output via pinMode() but accidentally calls pinMode() again when it looks like the intention was to initialize the chip select output to "high" level.  Looks like a copy and paste typo. That pin must default to a high level output so it worked, or was set to correct levels soon enough to not matter.